requirements.txt
# requirements.txt

asgiref==3.5.2
Django==3.2.12
django-cors-headers==3.13.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
PyJWT==2.6.0
pymongo==3.12.0
pytz==2022.2.1
sqlparse==0.2.4
    